requirements.txt

python-telegram-bot==20.7
ccxt
python-dotenv
